requirements.txt
matplotlib                3.2.2     
numpy                     1.19.1           
opencv                    4.0.1            
pillow                    7.2.0        
python                    3.8.5                
pytorch                   1.6.0      
scipy                     1.5.2      
sympy                     1.6.1      
tensorboard               2.3.0      





























